move session manager related files in separate dir
