+++ packages/web-sdk/src/config/makeCoreConfig.ts
@@ -11,8 +11,8 @@ import type { Config, MetaSession, Transport } from '@grafana/faro-core';
 import type { MetaItem } from '..';
 import { defaultEventDomain } from '../consts';
 import { parseStacktrace } from '../instrumentations';
-import { PersistentSessionsManager } from '../instrumentations/session/PersistentSessionsManager';
-import { isUserSessionValid } from '../instrumentations/session/sessionManagerUtils';
+import { PersistentSessionsManager } from '../instrumentations/session/sessionManager';
+import { isUserSessionValid } from '../instrumentations/session/sessionManager/sessionManagerUtils';
 import { createSession, defaultMetas, defaultViewMeta } from '../metas';
 import { k6Meta } from '../metas/k6';
 import { FetchTransport } from '../transports';
