+++ packages/web-sdk/src/instrumentations/session/index.ts
@@ -1,4 +1,3 @@
 export { SessionInstrumentation } from './instrumentation';
 
-export * from './types';
-export * from './VolatileSessionManager';
+export * from './sessionManager';
