+++ packages/web-sdk/src/instrumentations/session/sessionManager/PersistentSessionsManager.test.ts
