+++ packages/web-sdk/src/instrumentations/session/sessionManager/VolatileSessionsManager.test.ts
