+++ packages/web-sdk/src/instrumentations/session/sessionManager/index.ts
@@ -1,0 +1,3 @@
+export { PersistentSessionsManager } from './PersistentSessionsManager';
+export { VolatileSessionsManager } from './VolatileSessionManager';
+export { FaroUserSession } from './types';
