+++ packages/web-sdk/src/instrumentations/session/sessionManager/sessionManagerUtils.test.ts
