+++ packages/web-sdk/src/instrumentations/session/sessionManager/sessionManagerUtils.ts
