+++ packages/web-sdk/src/instrumentations/session/sessionManager/types.ts
